+++ integration_tests/suite/test_db_tenant.py
@@ -9,7 +9,6 @@ from hamcrest import (
     calling,
     contains,
     contains_inanyorder,
-    empty,
     equal_to,
     has_entries,
     has_properties,
@@ -32,48 +31,6 @@ def teardown():
 
 
 class TestTenantDAO(base.DAOTestCase):
-
-    @fixtures.tenant()
-    @fixtures.policy()
-    def test_add_policy(self, policy_uuid, tenant_uuid):
-        assert_that(self._policy_dao.list_(tenant_uuid=tenant_uuid), empty())
-
-        self._tenant_dao.add_policy(tenant_uuid, policy_uuid)
-        assert_that(
-            self._policy_dao.list_(tenant_uuid=tenant_uuid),
-            contains(has_entries(uuid=policy_uuid)),
-        )
-
-        self._tenant_dao.add_policy(tenant_uuid, policy_uuid)  # twice
-
-        assert_that(
-            calling(self._tenant_dao.add_policy).with_args(self.unknown_uuid, policy_uuid),
-            raises(exceptions.UnknownTenantException),
-            'unknown tenant',
-        )
-
-        assert_that(
-            calling(self._tenant_dao.add_policy).with_args(tenant_uuid, self.unknown_uuid),
-            raises(exceptions.UnknownPolicyException),
-            'unknown policy',
-        )
-
-    @fixtures.tenant()
-    @fixtures.policy()
-    def test_remove_policy(self, policy_uuid, tenant_uuid):
-        result = self._tenant_dao.remove_policy(tenant_uuid, policy_uuid)
-        assert_that(result, equal_to(0))
-
-        self._tenant_dao.add_policy(tenant_uuid, policy_uuid)
-
-        result = self._tenant_dao.remove_policy(self.unknown_uuid, policy_uuid)
-        assert_that(result, equal_to(0))
-
-        result = self._tenant_dao.remove_policy(tenant_uuid, self.unknown_uuid)
-        assert_that(result, equal_to(0))
-
-        result = self._tenant_dao.remove_policy(tenant_uuid, policy_uuid)
-        assert_that(result, equal_to(1))
 
     @fixtures.tenant(name='c')
     @fixtures.tenant(name='b')