+++ wazo_auth/plugins/http/tenant_policy/plugin.py
@@ -10,20 +10,9 @@ class Plugin(object):
     def load(self, dependencies):
         api = dependencies['api']
         tenant_service = dependencies['tenant_service']
-        policy_service = dependencies['policy_service']
 
-        api.add_resource(
-            http.TenantPolicy,
-            '/tenants/<uuid:tenant_uuid>/policies/<uuid:policy_uuid>',
-            resource_class_args=(tenant_service,)
-        )
         api.add_resource(
             http.TenantPolicies,
             '/tenants/<uuid:tenant_uuid>/policies',
             resource_class_args=(tenant_service,),
-        )
-        api.add_resource(
-            http.PolicyTenants,
-            '/policies/<uuid:policy_uuid>/tenants',
-            resource_class_args=(policy_service,),
         )